refactor(cli): use new deployment api
